Improved event management, improved timeline, fixed memory leaks, improved Osd

+++ trunk/ktoon/src/components/timeline/kttlruler.cpp
@@ -64,7 +64,7 @@ void KTTLRuler::paintSection ( QPainter * painter, const QRect & rect, int logic
 		
 		QString number = QString::number(logicalIndex+1);
 		
-		painter->drawText(rect.center().x()-(fm.width(number)/2), rect.center().y() +(fm.height()/2) ,number);
+		painter->drawText( (int)(rect.center().x()-(fm.width(number)/2)), (int)(rect.center().y() +(fm.height()/2)) ,number);
 	}
 	
 	painter->drawLine(rect.bottomLeft(), rect.bottomLeft()- QPointF(0, 4));
